Add files via upload
This was a special “one-off” report that one member of the Executive Committee wanted answered.  Based on major systems aboard certain vessels/divisions, where are the costs going to overall with EDL Delays during production.  I know that is a mouthful, but I actually had to have a meeting after not understanding the entirety of what was being asked, and kept submitting reports that did not answer his question, but I couldn’t grasp what he was actually asking.  So this is key, is to understand what the target audience (or key audience) needs to have answered.  Sometimes I create the questions/answers and present because my audience doesn’t know what they need to know yet, and the reports have and will evolve over time to answer those questions and others that come up during each meeting.  I had started to semi-automate it for if it did become a standard periodic report, but it was never asked for again.